Nodejs Project Wifibuddy
A web application that lists the WIFI connection information and password of a place with Nodejs. This Project aims to share the needed information about the cafe or another popular spot. When one person was going the place, they can instantly connect this system fort his application. This Project is working on Nodejs and MySQL.